requirements.txt
asgiref==3.7.2
asttokens==2.4.1
coverage==7.4.0
decorator==5.1.1
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.19.0
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
uritemplate==4.1.1
wcwidth==0.2.12
